+++ eras/conway/test-suite/src/Test/Cardano/Ledger/Conway/Serialisation/Roundtrip.hs
@@ -7,13 +7,15 @@
 
 module Test.Cardano.Ledger.Conway.Serialisation.Roundtrip (allprops) where
 
+import Cardano.Ledger.Alonzo.TxWits (AlonzoTxWits)
 import Cardano.Ledger.Binary.Version (natVersion)
+import Cardano.Ledger.Conway (Conway)
 import Cardano.Ledger.Conway.Genesis (ConwayGenesis (..))
-import Cardano.Ledger.Core (Era (..))
+import Cardano.Ledger.Core
 import Data.Data (Proxy (..), typeRep)
 import Test.Cardano.Ledger.Alonzo.Arbitrary (FlexibleCostModels)
 import Test.Cardano.Ledger.Binary.Plain.RoundTrip as Plain (roundTripCborExpectation)
-import Test.Cardano.Ledger.Binary.RoundTrip (roundTripCborRangeExpectation)
+import Test.Cardano.Ledger.Binary.RoundTrip
 import Test.Cardano.Ledger.Conway.Arbitrary ()
 import Test.Tasty (TestTree, testGroup)
 import Test.Tasty.QuickCheck (testProperty)
@@ -21,7 +23,7 @@ import Test.Tasty.QuickCheck (testProperty)
 allprops :: forall e. Era e => TestTree
 allprops =
   testGroup
-    (show $ typeRep (Proxy @e))
+    (show (typeRep (Proxy @e)) <> " CBOR")
     [ testProperty "ConwayGenesis" $
         roundTripCborRangeExpectation @(ConwayGenesis (EraCrypto e))
           (natVersion @2)
@@ -32,4 +34,16 @@ allprops =
         roundTripCborRangeExpectation @FlexibleCostModels
           (natVersion @9)
           (natVersion @9)
+    , testProperty "Conway Script" $
+        roundTripAnnRangeExpectation @(Script Conway)
+          (eraProtVerLow @Conway)
+          (eraProtVerHigh @Conway)
+    , testProperty "Conway TxWits" $
+        roundTripAnnRangeExpectation @(AlonzoTxWits Conway)
+          (eraProtVerLow @Conway)
+          (eraProtVerHigh @Conway)
+    , testProperty "Conway AuxiliaryData" $
+        roundTripAnnRangeExpectation @(TxAuxData Conway)
+          (eraProtVerLow @Conway)
+          (eraProtVerHigh @Conway)
     ]