+++ libs/cardano-ledger-api/test/Test/Cardano/Ledger/Api/Tx/Out.hs
@@ -17,8 +17,8 @@ import qualified Cardano.Ledger.Val as Val
 import qualified Data.ByteString.Lazy as BSL
 import Data.Functor.Identity
 import Lens.Micro
-import Test.Cardano.Ledger.Babbage.Arbitrary ()
 import Test.Cardano.Ledger.Common
+import Test.Cardano.Ledger.Conway.Arbitrary ()
 
 propSetShelleyMinTxOut ::
   forall era.