+++ eras/alonzo/impl/CHANGELOG.md
@@ -7,6 +7,8 @@
   * `DState`
   * `PState`
   * `VState`
+* Added support for Plutus V3 in the types and functions that use `Language`.
+  (Note that the Alonzo ledger era rules do not allow V3 scripts, however.).
 
 ## 1.1.0.0