+++ eras/conway/impl/CHANGELOG.md
@@ -15,6 +15,9 @@
 * Added `ConwayCERT` rule synonym
 * Added `ConwayDelegsPredFailure` rule
 * Added `ConwayDelegsEvent` rule
+* Change the Conway txInfo to allow Plutus V3
+  NOTE - unlike V1 and V2, the ledger will no longer place the "zero ada" value
+  in the script context for the transaction minting field.
 
 ## 1.1.0.0